5.x-feat-spid-test-configuration: refactor SigningCredential class
